refactor[wip]: Improves folder layout
