+++ src/on-chain-data/wt-hotel-index.js
@@ -1,10 +1,11 @@
 // @flow
 import type { WTHotelIndexInterface, HotelInterface, PreparedTransactionMetadataInterface } from '../interfaces/hotel-interfaces';
-import Utils from '../utils';
-import Contracts from '../contracts';
+import Utils from './utils';
+import Contracts from './contracts';
 import OnChainHotel from './on-chain-hotel';
 
-import { InputDataError, HotelNotFoundError, HotelNotInstantiableError, WTLibsError } from '../errors';
+import { InputDataError, HotelNotFoundError, HotelNotInstantiableError } from './errors';
+import { WTLibsError } from '../errors';
 import AbstractWTIndex from './wt-index';
 
 /**
