+++ test/wt-libs/on-chain-data/contracts.spec.js
@@ -1,7 +1,7 @@
 import { assert } from 'chai';
 import sinon from 'sinon';
-import Contracts from '../../src/contracts';
-import { SmartContractInstantiationError } from '../../src/errors';
+import Contracts from '../../../src/on-chain-data/contracts';
+import { SmartContractInstantiationError } from '../../../src/on-chain-data/errors';
 
 describe('WTLibs.Contracts', () => {
   let contracts, getCodeStub, ContractStub;
