+++ test/wt-libs/on-chain-data/hotels/hotel-model.spec.js
@@ -1,13 +1,13 @@
 import { assert } from 'chai';
 import sinon from 'sinon';
-import { AirlineDataModel } from '../../../../src/data-model/index';
-import WTAirlineIndex from '../../../../src/data-model/wt-airline-index';
-import testedDataModel from '../../../utils/data-airline-model-definition';
+import { HotelDataModel } from '../../../../src/on-chain-data/index';
+import WTHotelIndex from '../../../../src/on-chain-data/wt-hotel-index';
+import testedDataModel from '../../../utils/data-hotel-model-definition';
 
-describe('WTLibs.data-model.AirlineDataModel', () => {
+describe('WTLibs.on-chain-data.HotelDataModel', () => {
   it('should cache WTIndex instances', () => {
-    const dataModel = AirlineDataModel.createInstance(testedDataModel.withDataSource().dataModelOptions);
-    const createInstanceSpy = sinon.spy(WTAirlineIndex, 'createInstance');
+    const dataModel = HotelDataModel.createInstance(testedDataModel.withDataSource().dataModelOptions);
+    const createInstanceSpy = sinon.spy(WTHotelIndex, 'createInstance');
     assert.equal(createInstanceSpy.callCount, 0);
     dataModel.getWindingTreeIndex('address1');
     assert.equal(createInstanceSpy.callCount, 1);
