+++ test/wt-libs/on-chain-data/hotels/on-chain-hotel.spec.js
@@ -1,12 +1,12 @@
 import { assert } from 'chai';
 import sinon from 'sinon';
 import helpers from '../../../utils/helpers';
-import OnChainHotel from '../../../../src/data-model/on-chain-hotel';
-import StoragePointer from '../../../../src/storage-pointer';
+import OnChainHotel from '../../../../src/on-chain-data/on-chain-hotel';
+import StoragePointer from '../../../../src/on-chain-data/storage-pointer';
 
-import { InputDataError, SmartContractInstantiationError } from '../../../../src/errors';
+import { InputDataError, SmartContractInstantiationError } from '../../../../src/on-chain-data/errors';
 
-describe('WTLibs.data-model.OnChainHotel', () => {
+describe('WTLibs.on-chain-data.OnChainHotel', () => {
   let contractsStub, createdStub, utilsStub, indexContractStub, urlStub, managerStub;
   const validUri = 'schema://new-url';
   const validManager = 'manager';
