+++ test/wt-libs/on-chain-data/storage-pointer.spec.js
@@ -1,9 +1,10 @@
 import { assert } from 'chai';
 import sinon from 'sinon';
-import StoragePointer from '../../src/storage-pointer';
-import OffChainDataClient from '../../src/off-chain-data-client';
 import InMemoryAdapter from '@windingtree/off-chain-adapter-in-memory';
-import { StoragePointerError, OffChainDataRuntimeError } from '../../src/errors';
+import StoragePointer from '../../../src/on-chain-data/storage-pointer';
+import { StoragePointerError } from '../../../src/on-chain-data/errors';
+import { OffChainDataClient } from '../../../src/off-chain-data-client';
+import { OffChainDataRuntimeError } from '../../../src/off-chain-data-client/errors';
 
 describe('WTLibs.StoragePointer', () => {
   beforeEach(() => {
